Fix: Try to fix test cases
