+++ care/facility/tests/test_asset_public_api.py
@@ -1,11 +1,75 @@
+from enum import Enum
+
 from django.core.cache import cache
 from rest_framework import status
 from rest_framework.test import APIRequestFactory, APITestCase
+from rest_framework_simplejwt.tokens import RefreshToken
 
 from care.facility.api.viewsets.asset import AssetPublicViewSet
 from care.facility.models import Asset, AssetLocation
 from care.facility.tests.mixins import TestClassMixin
 from care.utils.tests.test_base import TestBase
+
+
+class ExpectedLocationObjectKeys(Enum):
+    ID = "id"
+    FACILITY = "facility"
+    CREATED_DATE = "created_date"
+    MODIFIED_DATE = "modified_date"
+    NAME = "name"
+    DESCRIPTION = "description"
+    LOCATION_TYPE = "location_type"
+
+
+class ExpectedFacilityKeys(Enum):
+    ID = "id"
+    NAME = "name"
+
+
+class ExpectedAssetRetrieveKeys(Enum):
+    ID = "id"
+    STATUS = "status"
+    ASSET_TYPE = "asset_type"
+    CREATED_DATE = "created_date"
+    MODIFIED_DATE = "modified_date"
+    NAME = "name"
+    DESCRIPTION = "description"
+    ASSET_CLASS = "asset_class"
+    IS_WORKING = "is_working"
+    NOT_WORKING_REASON = "not_working_reason"
+    SERIAL_NUMBER = "serial_number"
+    WARRANTY_DETAILS = "warranty_details"
+    META = "meta"
+    VENDOR_NAME = "vendor_name"
+    SUPPORT_NAME = "support_name"
+    SUPPORT_PHONE = "support_phone"
+    SUPPORT_EMAIL = "support_email"
+    QR_CODE_ID = "qr_code_id"
+    MANUFACTURER = "manufacturer"
+    WARRANTY_AMC_END_OF_VALIDITY = "warranty_amc_end_of_validity"
+    LAST_SERVICED_ON = "last_serviced_on"
+    NOTES = "notes"
+
+
+class ExpectedPublicAssetRetrieveKeys(Enum):
+    ID = "id"
+    LOCATION_OBJECT = "location_object"
+    EXTERNAL_ID = "external_id"
+    CREATED_DATE = "created_date"
+    MODIFIED_DATE = "modified_date"
+    DELETED = "deleted"
+    NAME = "name"
+    ASSET_TYPE = "asset_type"
+    STATUS = "status"
+    IS_WORKING = "is_working"
+    SERIAL_NUMBER = "serial_number"
+    WARRANTY_DETAILS = "warranty_details"
+    META = "meta"
+    VENDOR_NAME = "vendor_name"
+    SUPPORT_NAME = "support_name"
+    SUPPORT_PHONE = "support_phone"
+    SUPPORT_EMAIL = "support_email"
+    CURRENT_LOCATION = "current_location"
 
 
 class AssetPublicViewSetTestCase(TestBase, TestClassMixin, APITestCase):
@@ -15,6 +79,11 @@ class AssetPublicViewSetTestCase(TestBase, TestClassMixin, APITestCase):
         district = self.create_district(state=state)
         self.user = self.create_user(district=district, username="test user")
         facility = self.create_facility(district=district, user=self.user)
+        # Add access token to the authorization header of test request
+        refresh_token = RefreshToken.for_user(self.user)
+        self.client.credentials(
+            HTTP_AUTHORIZATION=f"Bearer {refresh_token.access_token}"
+        )
         self.asset_location = AssetLocation.objects.create(
             name="asset1 location", location_type=1, facility=facility
         )
@@ -23,14 +92,16 @@ class AssetPublicViewSetTestCase(TestBase, TestClassMixin, APITestCase):
         )
 
     def test_retrieve_asset(self):
-        response = self.new_request(
-            (f"/api/v1/public/asset/{self.asset.external_id}/",),
-            {"get": "retrieve"},
-            AssetPublicViewSet,
-            self.user,
-            {"external_id": str(self.asset.external_id)},
-        )
+        response = self.client.get(f"/api/v1/public/asset/{self.asset.external_id}/")
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        data = response.json()
+        expected_keys = [key.value for key in ExpectedPublicAssetRetrieveKeys]
+        self.assertCountEqual(data.keys(), expected_keys)
+        location_object_keys = [key.value for key in ExpectedLocationObjectKeys]
+        self.assertCountEqual(data["location_object"].keys(), location_object_keys)
+        facility_keys = [key.value for key in ExpectedFacilityKeys]
+        self.assertCountEqual(data["location_object"]["facility"].keys(), facility_keys)
 
     def test_retrieve_cached_asset(self):
         key = "asset:" + str(self.asset.external_id)
